Refactor MessageBus: - add service identity - add IMessageBus interface to avoid cyclic deps between modules (plugins and request handlers) - add Non-blocking handlers to support nested requests (see cancel handler) - add response validator to filter incoming responses - support multiple subscriptions on the same subject - return proper response error codes for timeout and expiration - support child request cancellation and wait on completion

+++ shared/javascript/core-lib/src/messaging/ResponseMatcher.ts
@@ -39,9 +39,12 @@ export class ResponseMatcher {
     match(msg: MessageEnvelope<Response>): Boolean {
         const ctx = this.requests.get(msg.body.requestId);
         if (ctx) {
-            this.receivedResponses.push(msg);
-            if (ctx.increment() >= (ctx.options.limit ?? 1)) {
-                this.requests.delete(msg.body.requestId);
+            const isValidResp = ctx.options.validator?.(msg.body) ?? true;
+            if (isValidResp) {
+                this.receivedResponses.push(msg);
+                if (ctx.increment() >= (ctx.options.limit ?? 1)) {
+                    this.requests.delete(msg.body.requestId);
+                }
             }
             return true;
         }
